+++ packages/core/styles/src/packages/composite/table/table.sprinkle.css.ts
@@ -1,0 +1,23 @@
+import {
+  createRainbowSprinkles,
+  defineProperties as defineRainbowProperties,
+} from "rainbow-sprinkles";
+import { paddingProperties } from "../../../properties";
+
+const defineProperties = defineRainbowProperties({
+  dynamicProperties: {
+    width: true,
+    maxWidth: true,
+    minWidth: true,
+    padding: paddingProperties,
+  },
+});
+
+const sprinkle = createRainbowSprinkles(defineProperties);
+
+export const tableSprinkle = {
+  sprinkle,
+  properties: {
+    padding: paddingProperties,
+  },
+};