+++ packages/react/src/composite/Table/src/components/TableBody/TableBody.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+import { table } from "@nimbus-ds/styles";
+
+import { TableBodyProps } from "./tableBody.types";
+
+const TableBody: React.FC<TableBodyProps> = ({
+  className: _className,
+  style: _style,
+  children,
+  ...rest
+}: TableBodyProps) => (
+    <tbody {...rest} className={table.classnames.body}>
+      {children}
+    </tbody>
+  );
+
+export { TableBody };