+++ packages/react/src/composite/Table/src/components/TableBody/index.ts
@@ -1,0 +1,4 @@
+import { TableBody } from "./TableBody";
+
+export { TableBody } from "./TableBody";
+export default TableBody;