+++ packages/react/src/composite/Table/src/components/TableBody/tableBody.types.ts
@@ -1,0 +1,6 @@
+import { ReactNode, HTMLAttributes } from "react";
+
+export interface TableBodyProps
+  extends HTMLAttributes<HTMLTableSectionElement> {
+  children: ReactNode;
+}