+++ packages/react/src/composite/Table/src/components/TableCell/index.ts
@@ -1,0 +1,4 @@
+import { TableCell } from "./TableCell";
+
+export { TableCell } from "./TableCell";
+export default TableCell;