+++ packages/react/src/composite/Table/src/components/TableCell/tableCell.types.ts
@@ -1,0 +1,8 @@
+import { ReactNode, HTMLAttributes } from "react";
+import { TableSprinkle } from "@nimbus-ds/styles";
+
+type TableExtends = TableSprinkle & HTMLAttributes<HTMLTableCellElement>;
+
+export interface TableCellProps extends TableExtends {
+  children: ReactNode;
+}