+++ packages/react/src/composite/Table/src/components/TableHead/tableHead.types.ts
@@ -1,0 +1,6 @@
+import { ReactNode, HTMLAttributes } from "react";
+
+export interface TableHeadProps
+  extends HTMLAttributes<HTMLTableSectionElement> {
+  children: ReactNode;
+}