+++ packages/react/src/composite/Table/src/index.ts
@@ -1,0 +1,5 @@
+import { Table } from "./Table";
+
+export { Table } from "./Table";
+export type { TableProps } from "./table.types";
+export default Table;