+++ packages/react/src/composite/Table/src/table.spec.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+import { render, screen } from "@testing-library/react";
+
+import { Table } from "./Table";
+import { TableProps } from "./table.types";
+
+const makeSut = (rest: TableProps) => {
+  render(<Table {...rest} data-testid="table-element" />);
+};
+
+describe("GIVEN <Table />", () => {
+  describe("WHEN rendered", () => {
+    it("THEN should correctly render the submitted content", () => {
+      makeSut({ children: <div>My content</div> });
+      expect(screen.getByText("My content")).toBeDefined();
+    });
+  });
+});